+++ ChangeLog.md
@@ -4,6 +4,7 @@
 * Allow type inference in initializer of `thread_local!`.
 * Add `with` implementation that panics if neither `host` nor `plugin` features
   are enabled.
+* Allow both `host` and `plugin` to be enabled together.
 
 # 0.1.0  -- 2020-12-01
 * Initial release.